requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
httpx[http2]
ffmpeg-python
gitpython
hachoir
heroku3
motor
pillow
psutil
py-tgcalls==0.9.7
pykeyboard
git+https://github.com/KurimuzonAkuma/pyrogram@dev
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
TgCrypto-pyrofork
unidecode
yt-dlp
youtube-search
youtube-search-python
